[feature:frontend/product-hooks] - new refactoring product components and hooks
