+++ client/src/app/product/index.tsx
@@ -1,15 +1,15 @@
 import { useState } from "react";
-import { ProductsForm } from "@/app/product/product-form";
-import { ProductsTable } from "@/app/product/product-table";
+import { ProductsForm } from "@/app/product/components/product-form";
+import { ProductsTable } from "@/app/product/components/product-table";
 import { FunnelSimple, Package } from "@phosphor-icons/react";
-import { ProductsFilter } from "@/app/product/product-filter";
+import { ProductsFilter } from "@/app/product/components/product-filter";
 import { MODALS } from "@/components/modal/types";
 import { ModalOpenButton } from "@/components/modal/modal-open-button";
-import { ProductDelete } from "@/app/product/product-delete";
+import { ProductDelete } from "@/app/product/components/product-delete";
 import { useSelectedProductStore } from "./product.store";
 
-export function ProductsList() {
-  const { setSelectedProduct } = useSelectedProductStore();
+export function ProductsIndex() {
+  const { setProduct } = useSelectedProductStore();
   const [filterIsOpen, setFilterIsOpen] = useState<boolean>(false);
 
   return (
@@ -23,7 +23,7 @@ export function ProductsList() {
             <ModalOpenButton
               type="button"
               targetId={MODALS.PRODUCT_FORM}
-              onClick={() => setSelectedProduct(undefined)}
+              onClick={() => setProduct(undefined)}
               className="btn btn-dark d-flex gap-3 align-items-center"
             >
               Adicionar Produto {<Package size={32} />}
